+++ config.js
@@ -1,47 +1,47 @@
 export default {
-    prefix: '/api',
-    port: 3000,
+    prefix: "/api",
+    port: 3001,
     verbs: {
         get: {
-            findUnique: '/:id',
-            findMany: '',
-            findFirst: '/first',
-            count: '/count',
-            aggregate: '/aggregate',
-            groupBy: '/groupBy',
+            findUnique: "/:id",
+            findMany: "",
+            findFirst: "/first",
+            count: "/count",
+            aggregate: "/aggregate",
+            groupBy: "/groupBy",
         },
         post: {
-            create: '',
+            create: "",
         },
         put: {
-            upsert: '',
+            upsert: "",
         },
         patch: {
-            update: '/:id',
-            updateMany: '/update',
+            update: "/:id",
+            updateMany: "/update",
         },
         delete: {
-            delete: '/:id',
-            deleteMany: '',
-        }
+            delete: "/:id",
+            deleteMany: "",
+        },
     },
     tokenVerify: true,
-    tokenKey: 'EXPOSER_TOKEN_KEY',
+    tokenKey: "EXPOSER_TOKEN_KEY",
     aclVerify: true,
-    aclType: 'fast',
+    aclType: "fast",
     userModel: {
-        roleId: 'roleId',
-        defaultRoleId: 1
+        roleId: "roleId",
+        defaultRoleId: 1,
     },
     roleModel: {
-        tableName: 'role',
-        id: 'id',
-        name: 'name'
+        tableName: "role",
+        id: "id",
+        name: "name",
     },
     aclModel: {
-        tableName: 'acls',
-        model: 'model',
-        name: 'name',
-        allow: 'allow',
-    }
-}
+        tableName: "acls",
+        model: "model",
+        name: "name",
+        allow: "allow",
+    },
+};