+++ index.js
@@ -8,11 +8,13 @@ import hooks from './src/hooks.js';
 
 //Middlewares
 import notFound from './src/middleware/notFound.js';
+import tokenVerification from './src/middleware/tokenVerification.js';
+
 import UserError from './src/errors/userError.js';
 import { addModel } from './src/acl/aclVerification.js';
-import tokenVerification from './src/middleware/tokenVerification.js';
-import useUser from './src/user/useUser.js';
+import autoImport from './src/util/autoImport.js';
 import useAcl from './src/acl/useAcl.js';
+import useUser from './src/user/useUser.js';
 
 //Express
 import express, { json, urlencoded } from 'express';
@@ -21,20 +23,18 @@ import cors from 'cors';
 //Default ORM [Prisma]
 import { exporter } from 'exposerjs-orm-prisma';
 
-async function run(prismaClient, app, userConfig) {
-    if (!prismaClient) throw new Error('Prisma is required!');
-
+async function run({ config: userConfig, PrismaClient, app }) {
     const st = new Date();
-    const hasExpress = app ? true : false
+    const notHasExpress = !app
     global.CONFIG = Object.assign(config, userConfig);
-    global.ORM = exporter(prismaClient)
+    global.ORM = exporter(PrismaClient) //await
 
-    if (!hasExpress) {
-        app = express()
-        app.use(cors());
-        app.use(json());
-        app.use(urlencoded({ extended: true }));
+    if (userConfig.autoImport) {
+        const { paths, getFile } = userConfig.autoImport
+        await autoImport(paths, getFile)
     }
+    if (notHasExpress) app = express()
+
 
     useUser();
     app.use(tokenVerification);
@@ -47,13 +47,19 @@ async function run(prismaClient, app, userConfig) {
     app.use(notFound);
 
     console.log(`ExposerJS deployed in ${new Date() - st}ms ⚡`.bgCyan);
-    if (!hasExpress) {
-        return app.listen(global.CONFIG.port, () => {
-            console.log(`BackEnd is listen at port ${global.CONFIG.port}`.bgGreen);
-            return app
-        });
-    }
+    if (notHasExpress) return expressDeploy(app)
 };
+
+function expressDeploy(app) {
+    app.use(cors());
+    app.use(json());
+    app.use(urlencoded({ extended: true }));
+    return app.listen(global.CONFIG.port, () => {
+        console.log(`BackEnd is listen at port ${global.CONFIG.port}`.bgGreen);
+        return app
+    });
+}
+
 
 const exposer = { use, run, UserError }
 const acl = { addModel }